+++ packages/expo-ui/android/src/main/java/expo/modules/ui/Utils.kt
@@ -4,6 +4,8 @@ import android.graphics.Color
 import android.os.Build
 import android.util.Log
 import androidx.compose.ui.graphics.vector.ImageVector
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
 
 fun colorToComposeColorOrNull(color: Color?): androidx.compose.ui.graphics.Color? {
   return color?.let {
@@ -39,3 +41,7 @@ fun getImageVector(icon: String?): ImageVector? {
     return null
   }
 }
+
+data class GenericEventPayload1<T>(
+  @Field val value: T
+) : Record